updater

+++ src/WinInstaller.Uninstaller/ScriptInterop.cs
@@ -1,4 +1,5 @@
 ﻿using System.IO;
+using System.Windows;
 using WinInstaller.Uninstaller.Engine;
 
 namespace WinInstaller.Uninstaller;
@@ -9,23 +10,35 @@ public class ScriptInterop : ScriptInteropBase
 
     public void UnInstall()
     {
-        App.CurrentInstance.Running = true;
-        App.CurrentInstance.WebBrowser.InvokeScript("setRunningState", true);
-        //1.删除文件
-        DeleteFolder(App.CurrentInstance.Config.InstallLocation);
+        if (MessageBox.Show("卸载前请注意备份文件,确定卸载?", "卸载确认", MessageBoxButton.YesNo) == MessageBoxResult.Yes)
+        {
+            try
+            {
+                App.CurrentInstance.Running = true;
+                App.CurrentInstance.WebBrowser.InvokeScript("setRunningState", true);
+                //1.删除文件
+                DeleteFolder(App.CurrentInstance.Config.InstallLocation);
 
-        //2.删除注册表
-        App.CurrentInstance.WebBrowser.InvokeScript("setLog", $"删除注册表");
-        App.CurrentInstance.Config.Delete();
+                //2.删除注册表
+                App.CurrentInstance.WebBrowser.InvokeScript("setLog", $"删除注册表");
+                App.CurrentInstance.Config.Delete();
 
-        //3.删除快捷方式
-        var name = $"{App.CurrentInstance.Config.DisplayName}.lnk";
-        var desktopPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.DesktopDirectory), name);
-        var userStartPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.StartMenu), name);
-        App.CurrentInstance.WebBrowser.InvokeScript("setLog", $"删除快捷方式");
-        if (File.Exists(desktopPath)) File.Delete(desktopPath);
-        if (File.Exists(userStartPath)) File.Delete(userStartPath);
-        App.CurrentInstance.WebBrowser.InvokeScript("setRunningState", true);
+                //3.删除快捷方式
+                var name = $"{App.CurrentInstance.Config.DisplayName}.lnk";
+                var desktopPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.DesktopDirectory), name);
+                var userStartPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.StartMenu), name);
+                App.CurrentInstance.WebBrowser.InvokeScript("setLog", $"删除快捷方式");
+                if (File.Exists(desktopPath)) File.Delete(desktopPath);
+                if (File.Exists(userStartPath)) File.Delete(userStartPath);
+                App.CurrentInstance.WebBrowser.InvokeScript("setSuccessState", true);
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message);
+            }
+        }
+
+        App.CurrentInstance.WebBrowser.InvokeScript("setRunningState", false);
         App.CurrentInstance.Running = false;
     }
